реализует возможность автовоспроизведения првеью карточки фильма при наведении курсора
